Rearrange functions with and without hyperparameter optimization
