+++ requirements-hyperopt.txt
@@ -1,0 +1,3 @@
+torch-cluster
+ray[tune]
+hyperopt
