+++ requirements-torchdep.txt
@@ -1,4 +1,3 @@
-torch-cluster
 torch-geometric
 torch-scatter
 torch-sparse
