+++ requirements.txt
@@ -1,5 +1,4 @@
 matplotlib
 torch
 tqdm
-ray[tune]
-hyperopt
+tensorboard
